strstr1

+++ ft_strnew.c
@@ -16,7 +16,7 @@ char	*ft_strnew(size_t size)
 {
 	char *str;
 
-	str = (char *)malloc((int)size + 1);
+	str = (char *)malloc(size + 1);
 	str[size] = '\0';
 	return (str);
 }
